creacion de microservicio viaje y refactor

+++ monopatin-microservicio/src/main/java/com/monopatinmicroservicio/service/DTO/ReporteEstadoMonopatinesDTO.java
@@ -11,8 +11,8 @@ import java.io.Serializable;
 @AllArgsConstructor
 @NoArgsConstructor
 public class ReporteEstadoMonopatinesDTO implements Serializable {
-    private long aptos_para_uso;
-    private long en_mantenimiento;
+    private int aptos_para_uso;
+    private int en_mantenimiento;
